third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-localDescription-promise.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests RTCPeerConnection localDescription.");

var pc = null;

function requestFailed2()
{
    testPassed('requestFailed was called.');

    shouldBeEqualToString('pc.localDescription.type', "offer");
    shouldBeEqualToString('pc.localDescription.sdp', "local");
    pc.close();
    shouldBeEqualToString('pc.localDescription.type', "offer");
    shouldBeEqualToString('pc.localDescription.sdp', "local");

    finishJSTest();
}

function requestSucceeded2()
{
    testFailed('requestSucceeded was called.');
    finishJSTest();
}

function requestFailed1()
{
    testFailed('requestFailed was called.');
    finishJSTest();
}

function requestSucceeded1()
{
    testPassed('requestSucceeded was called.');

    sessionDescription = new RTCSessionDescription({type:"answer", sdp:"remote"});
    shouldNotThrow('pc.setLocalDescription(sessionDescription).then(requestSucceeded2, requestFailed2);');
}

function expectedTypeError(error)
{
    window.error = error;
    shouldBe('error.name', '"TypeError"')
    testPassed('expectedTypeError was called.')
}

function expectedInvalidSessionDescription(error)
{
    window.error = error;
    shouldBe('error.name', '"OperationError"')
    testPassed('expectedInvalidSessionDescription was called.')
}

pc = new RTCPeerConnection();
shouldNotThrow('pc.setLocalDescription().catch(expectedTypeError)');
shouldNotThrow('pc.setLocalDescription(null).catch(expectedInvalidSessionDescription)');
var sessionDescription = new RTCSessionDescription({type:"offer", sdp:"local"});
shouldNotThrow('pc.setLocalDescription(sessionDescription).then(requestSucceeded1, requestFailed1);');

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
</body>
</html>
