third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-localDescription-promise-expected.txt

Tests RTCPeerConnection localDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.setLocalDescription().catch(expectedTypeError) did not throw exception.
PASS pc.setLocalDescription(null).catch(expectedInvalidSessionDescription) did not throw exception.
PASS pc.setLocalDescription(sessionDescription).then(requestSucceeded1, requestFailed1); did not throw exception.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS error.name is "OperationError"
PASS expectedInvalidSessionDescription was called.
PASS requestSucceeded was called.
PASS pc.setLocalDescription(sessionDescription).then(requestSucceeded2, requestFailed2); did not throw exception.
PASS requestFailed was called.
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS successfullyParsed is true

TEST COMPLETE
